# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DailyStatPoint": {
        "properties": {
          "downloads": {
            "title": "Downloads",
            "type": "integer"
          },
          "stat_date": {
            "title": "Stat Date",
            "type": "string"
          },
          "views": {
            "title": "Views",
            "type": "integer"
          }
        },
        "required": [
          "stat_date",
          "views",
          "downloads"
        ],
        "title": "DailyStatPoint",
        "type": "object"
      },
...
      "DownloadTicketResponse": {
        "description": "docs/03 §3.15。",
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Name"
          },
          "file_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Sha256"
          },
          "file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "DownloadTicketResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SkillPreviewResponse": {
        "description": "docs/03 §3.5 的响应。",
        "properties": {
          "file_tree": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "File Tree",
            "type": "array"
          },
          "file_tree_truncated": {
            "default": false,
            "title": "File Tree Truncated",
            "type": "boolean"
          },
          "manifest": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest"
          },
          "readme_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readme Md"
          },
          "total_size": {
            "default": 0,
            "title": "Total Size",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version"
        ],
        "title": "SkillPreviewResponse",
        "type": "object"
      },
...
      "ToolFacets": {
        "description": "分类与类型的计数**已按当前用户可见性过滤**（FR-TAX-07）。\n\n**只在 `page == 1` 时返回**，翻页时整个 `facets` 键省略（docs/03 §3.3）。",
        "properties": {
          "categories": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CategoryFacet"
            },
            "title": "Categories",
            "type": "array"
          },
          "types": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TypeFacet"
            },
            "title": "Types",
            "type": "array"
          }
        },
        "title": "ToolFacets",
        "type": "object"
      },
...
      "ToolListResponse": {
        "properties": {
          "facets": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolFacets"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ToolListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "ToolListResponse",
        "type": "object"
      },
...
      "ToolStatsResponse": {
        "properties": {
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DailyStatPoint"
            },
            "title": "Daily",
            "type": "array"
          },
          "download_count": {
            "title": "Download Count",
            "type": "integer"
          },
          "tool_id": {
            "title": "Tool Id",
            "type": "integer"
          },
          "view_count": {
            "title": "View Count",
            "type": "integer"
          }
        },
        "required": [
          "tool_id",
          "download_count",
          "view_count"
        ],
        "title": "ToolStatsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "selftool — 内网工具 / Skill 共享平台后端",
    "title": "selftool API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tools": {
      "get": {
        "description": "门户主列表。\n\n几个要点：\n\n- **`sort` 白名单校验**，非法值返回 400 `INVALID_SORT` 而不是 500；\n  排序表达式从 `SORT_EXPRESSIONS` 查表得到，**绝不把 `sort` 拼进 SQL**\n  （docs/03 §1.6）。\n- **`facets` 只在 `page == 1` 时返回**，翻页时整个键省略（docs/03 §3.3）。\n- 可见性判定下推到 SQL（EXISTS 子查询），不在 Python 里过滤\n  （docs/01 §4.3 性能要求）。\n- 未登录用户由 `portal.allow_anonymous_view` 决定 401 还是返回公开数据\n  （契约 §6），该判断在 `portal_access` 依赖里完成。",
        "operationId": "list_tools_api_v1_tools_get",
        "parameters": [
          {
            "description": "关键词，搜索名称、简介、详情、标签",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "关键词，搜索名称、简介、详情、标签",
              "title": "Q"
            }
          },
          {
            "description": "分类 slug，可重复传",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类 slug，可重复传",
              "title": "Category"
            }
          },
          {
            "description": "标签名，可重复传",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "标签名，可重复传",
              "title": "Tag"
            }
          },
          {
            "description": "file/webapp/skill/prompt，可重复传",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "file/webapp/skill/prompt，可重复传",
              "title": "Type"
            }
          },
          {
            "description": "按作者 username 筛选",
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按作者 username 筛选",
              "title": "Owner"
            }
          },
          {
            "description": "hot（默认）/ new / name / -updated_at",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "hot（默认）/ new / name / -updated_at",
              "title": "Sort"
            }
          },
          {
            "description": "页码，从 1 开始",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "页码，从 1 开始",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "每页条数，1~200",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "每页条数，1~200",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "门户工具列表",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/tools/{slug}": {
      "get": {
        "description": "无权访问返回 **404**（不是 403）—— FR-FILE-08，避免探测资源存在性。\n\n浏览量在这里累加（内存计数器，不写库）。",
        "operationId": "get_tool_detail_api_v1_tools__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "工具详情",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/tools/{slug}/download": {
      "get": {
        "description": "走 ASGI `FileResponse`（sendfile 路径）。\n\n鉴权两条路径：\n  - `?ticket=` 有效票据 → 不需要 Authorization 头\n  - 否则常规凭证；匿名再看 `portal.allow_anonymous_view`\n\n`FileResponse` 自带 Range 支持（返回 206 + `Content-Range`），\n所以断点续传不用自己实现。",
        "operationId": "download_tool_api_v1_tools__slug__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "指定版本；不传则下载当前版本",
            "in": "query",
            "name": "version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "指定版本；不传则下载当前版本",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "下载（支持 Range 断点续传）",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/tools/{slug}/download-ticket": {
      "post": {
        "description": "签发 60 秒有效的 HMAC 票据（docs/03 §3.15）。\n\n票据绑定 `user_id`：泄露给别人也用不了。允许在有效期内多次使用，\n因为浏览器会为 Range 请求或重试发多次请求。",
        "operationId": "create_download_ticket_endpoint_api_v1_tools__slug__download_ticket_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "签发下载票据",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/tools/{slug}/stats": {
      "get": {
        "operationId": "tool_stats_api_v1_tools__slug__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "工具统计",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/tools/{slug}/versions": {
      "get": {
        "description": "对**有可见性权限的用户**开放（FR-VER-06）。\n\n`purged` 版本仍然出现在列表里但 `can_download=false` —— 界面显示「已归档」，\n而不是把它藏起来（那样用户会以为版本丢了）。",
        "operationId": "list_tool_versions_api_v1_tools__slug__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionSummary"
                  },
                  "title": "Response List Tool Versions Api V1 Tools  Slug  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "版本列表（含待审、已驳回、已归档）",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/tools/{slug}/versions/{version}/skill-preview": {
      "get": {
        "operationId": "skill_preview_api_v1_tools__slug__versions__version__skill_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Skill 包预览（manifest / readme / 文件树）",
        "tags": [
          "portal"
        ]
      }
    },
...
  }
}